# backend-kotlin/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: Spring Boot Kotlin Todo Application API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/todos:
    get:
      summary: Get all todos
      operationId: getAllTodos
      tags:
        - todos
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'

    post:
      summary: Create a new todo
      operationId: createTodo
      tags:
        - todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/todos/{id}:
    get:
      summary: Get todo by ID
      operationId: getTodoById
      tags:
        - todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update a todo
      operationId: updateTodo
      tags:
        - todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a todo
      operationId: deleteTodo
      tags:
        - todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Todo deleted successfully
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - task
        - completed
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 1
        task:
          type: string
          example: "Buy groceries"
        completed:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2025-01-09T10:30:00"

    TodoRequest:
      type: object
      required:
        - task
      properties:
        task:
          type: string
          minLength: 1
          example: "Buy groceries"
        completed:
          type: boolean
          example: false

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Todo not found"
        timestamp:
          type: string
          format: date-time