# backend-kotlin/src/main/kotlin/com/example/todo/model/Todo.kt
package com.example.todo.model

import jakarta.persistence.*
import jakarta.validation.constraints.NotBlank
import java.time.LocalDateTime

@Entity
@Table(name = "todos")
data class Todo(
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    val id: Long? = null,

    @field:NotBlank(message = "タスクは必須です")
    @Column(nullable = false)
    var task: String = "",

    @Column(nullable = false)
    var completed: Boolean = false,

    @Column(nullable = false, updatable = false)
    val createdAt: LocalDateTime = LocalDateTime.now()
)